perception/requirements.txt
fastapi
uvicorn
python-dotenv
slowapi
prometheus-fastapi-instrumentator
# If you plan to do actual CV or audio:
# opencv-python-headless
# librosa
